Make Executor TailContainer before RunContainer
The Kubernetes runtime needs to start asking for logs before the container is created,
or logs from short-lived steps might be gone by the time TailContainer starts.

The Docker runtime, however, needs to RunContainer first or the docker engine will
complain that the container does not exist. So, we use a special `runCtx` to let the
Docker Runtime's TailContainer know when RunContainer has finished to allow it
to delay pulling logs until the container exists.