refactor(k8s): Use PodTracker in WaitContainer
Replaces manual Pods Watch API call with a "container is terminated" signal
in a containerTracker. That signal is controlled by the PodTracker based
on Add/Update/Delete pod events from the PodInformer (which does its own
watch/list API calls internally).

NOTE: podTracker.Containers should have containerTracker pointers
the containerTracker has locks in it, so we don't
want to copy it. We want to get references to it.