chore: drop old comment and unused newMockWithWatch
Under ContainerStatus, ContainerStateTerminated must be nil when not terminated.
If terminated, that field must not be nil.

The only field on ContainerStateTerminated that is required,
and therefore required to be set is ExitCode.
So, waiting for Reason, an optional field, to be set sites not match the k8s API spec.

So, just drop the comment about maybe nursing this later.