refactor(k8s): move test logic to MockKubernetesRuntime interface
When we add kubernetes-based tests to the executor(s), we will need to enough access
to simulate certain kubernetes events. The MockKubernetesRuntime interface provides
access to all of those mocking methods.

This only introduces the methods that can be immediately used in k8s tests.
More mock utilities will be added in follow-ups.